docs(i2c-scanner): refactor the example
- Renamed from `i2c_scanner` to `i2c-scanner`.
- Simplified the I2C driver instantiation and usage.
- Formatted the output address as hexadecimal.
