+++ examples/i2c-scanner/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "i2c_scanner"
+name = "i2c-scanner"
 license.workspace = true
 edition.workspace = true
 publish = false
@@ -8,10 +8,6 @@ publish = false
 workspace = true
 
 [dependencies]
-cfg-if = { workspace = true }
-embassy-sync = { workspace = true }
-embedded-hal = { workspace = true }
-embedded-hal-async = { workspace = true }
-once_cell = { workspace = true }
 ariel-os = { path = "../../src/ariel-os", features = ["i2c"] }
 ariel-os-boards = { path = "../../src/ariel-os-boards" }
+embedded-hal-async = { workspace = true }
