+++ examples/i2c-scanner/src/pins.rs
