+++ examples/laze.yml
@@ -15,7 +15,7 @@ subdirs:
   - hello-world-threading
   - http-client
   - http-server
-  - i2c_scanner
+  - i2c-scanner
   - log
   - minimal
   - power
